pyproject.toml
[project]
name = "primatology"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.65.0",
    "openai>=2.24.0",
    "python-dotenv>=1.2.1",
    "rich>=14.3.3",
]
